requirements.txt
streamlit==1.22.0
nltk==3.8.1
pandas==2.0.0
plotly==5.14.1
finnhub-python==2.4.16
python-dotenv==1.0.0 